Allow file indirection of a session file
Capturing the output of `record session` into a file and using a value
is an idiom, except `--session $(cat session.txt)` is clunky and hostile
to Windows.

When a CI system is involved in the middle to pass around values like
this, sometimes it's easier to pass around session as an environment
variable. In such a case, receiving literal `--session $FOO` is more
convenient.

Furthermore, these two modes need to be both discoverable to users.

So I decided that the best course of action is to have one option,
`--session` that can deal with both, but by using a reasonably
well-established convention, that is `@path/to/file`.

This is called the response file, and fairly widespread on Windows. On
Unix, this is less common, but it's been used by tools like curl.