[chore] error reporting improvement
Tell apart a bad value vs no value. Also, use the ensure method
to avoid duplicating errors in multiple places.

Seeing incorrect value in LAUNCHABLE_TOKEN previously was a silently
recovered error, but I think this behavior is harmful. If the value is
incorrectly set, we need to call an attention to that fact so that we
can guide users to fix it.